feat(controller) : api 및 view controller 분리.
