+++ src/main/java/com/ovg/flipper/controller/api/AuthApiController.java
@@ -1,4 +1,4 @@
-package com.ovg.flipper.controller;
+package com.ovg.flipper.controller.api;
 
 import com.ovg.flipper.dto.UserAuthDto;
 import com.ovg.flipper.dto.UserLoginDto;
@@ -7,28 +7,22 @@ import com.ovg.flipper.service.UserAuthService;
 import com.ovg.flipper.util.CookieManager;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
-@Slf4j
-public class UserAuthController {
-
+@RestController
+@RequestMapping("/api")
+public class AuthApiController {
     private final UserAuthService userAuthService;
     private final CookieManager cookieManager;
 
-    public UserAuthController(UserAuthService userAuthService, CookieManager cookieManager) {
+    @Autowired
+    public AuthApiController(UserAuthService userAuthService, CookieManager cookieManager) {
         this.userAuthService = userAuthService;
         this.cookieManager = cookieManager;
-    }
-
-    @GetMapping("/login")
-    public String showLoginPage() {
-        return "login";
     }
 
     @PostMapping("/login")
@@ -36,6 +30,7 @@ public class UserAuthController {
         if (bindingResult.hasErrors()) {
             return "login";
         }
+
         UserAuthDto userAuthDto = userAuthService.login(user);
         if (userAuthDto == null) {
             return "redirect:/login?error";
@@ -47,31 +42,16 @@ public class UserAuthController {
         return "redirect:/mypage";
     }
 
-    @GetMapping("/signup")
-    public String showSignUpForm(Model model) {
-        model.addAttribute("user", new UserSignupDto());
-        return "signup";
-    }
-
     @PostMapping("/signup")
     public String signUp(@Valid UserSignupDto user, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             return "signup";
         }
+
         if(userAuthService.registerUser(user)){
             return "redirect:/login";
         }
+
         return "redirect:/signup?error";
-    }
-
-    // TEST : admin page
-    @GetMapping("/admin")
-    public String showAdminPage() {
-        return "admin";
-    }
-
-    @GetMapping("/403")
-    public String accessDenied() {
-        return "403";
     }
 }
