+++ src/main/resources/templates/login.html
@@ -6,7 +6,7 @@
 <body>
 <h1>Login</h1>
 
-<form th:action="@{/login}" method="post">
+<form th:action="@{/api/login}" method="post">
     <div>
         <label for="username">Username</label>
         <input type="text" id="username" name="username" placeholder="Enter your username"/>
