+++ src/main/resources/templates/signup.html
@@ -6,7 +6,7 @@
 <body>
 <h1>Sign Up</h1>
 
-<form action="#" th:action="@{/signup}" th:object="${user}" method="post">
+<form action="#" th:action="@{/api/signup}" th:object="${user}" method="post">
     <div>
         <label for="username">Username</label>
         <input type="text" id="username" th:field="*{username}" placeholder="Enter your username"/>
